railway/permissions.py
from rest_framework.permissions import SAFE_METHODS, BasePermission


class IsAdminOrIfAuthenticatedReadOnly(BasePermission):
    def has_permission(self, request, view):
        user = request.user

        if user and user.is_staff:
            return True
        if request.method in SAFE_METHODS and user and user.is_authenticated:
            return True

        return False
